requirements.txt
APScheduler==3.10.1
bcrypt==4.0.1
bidict==0.22.1
cachelib==0.10.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Session==0.4.0
Flask-SocketIO==5.3.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pymongo==4.3.3
python-dotenv==1.0.0
python-engineio==4.4.0
python-socketio==5.7.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.15
Werkzeug==2.2.3
